Refactor buttons
