--- src/components/buttons/ResetFiltersButton.jsx
@@ -1,22 +1,0 @@
-import React, { PropTypes } from 'react';
-import { RaisedButton } from 'material-ui';
-
-const labelStyle = {
-  fontSize: '12px',
-};
-
-const ResetFiltersButton = ({resetCallback}) => (
-  <RaisedButton
-    fullWidth
-    label="Reset"
-    labelStyle={labelStyle}
-    onTouchTap={resetCallback}
-    secondary
-  />
-);
-
-ResetFiltersButton.propTypes = {
-  resetCallback: PropTypes.func.isRequired,
-};
-
-export default ResetFiltersButton;
